Added animation to JS backend
