+++ js/src/main/scala/Main.scala
@@ -1,11 +1,14 @@
+import doodle.core._
+import doodle.examples._
 
 import scala.scalajs.js.JSApp
 import scala.scalajs.js.annotation.JSExport
 import org.scalajs.dom
 
 @JSExport object Main extends JSApp {
-  val image = doodle.examples.ChristmasTree
+  val image = ChristmasTree
+  val anim  = Orbit(Angle.radians(0), 50, image)
 
   @JSExport def main(): Unit =
-    doodle.js.draw(image, "canvas")
+    doodle.js.animate(anim, "canvas")
 }
